+++ tests/testthat/test-file_loaders.R
@@ -71,13 +71,38 @@ test_that("load_file with an .h5ad works", {
 ###################### TODO: add anndata_zarr tests ######################
 
 test_that("load_file with a .parquet works", {
-  skip_if_not_installed("nanoparquet")
+  skip_if_not_installed("arrow")
 
   file <- withr::local_file(tempfile(fileext = ".parquet"))
 
   # create a Parquet file
   df <- data.frame(a = 1:3, b = 4:6)
-  nanoparquet::write_parquet(df, file)
+  arrow::write_parquet(df, file)
+
+  # load the Parquet file
+  loaded_df <- load_file(file)
+
+  # ignore class differences
+  class(loaded_df) <- class(df)
+
+  # check that the data frame is the same
+  expect_equal(loaded_df, df)
+})
+
+test_that("reading/writing small_dataset1 as a .parquet works", {
+  skip_if_not_installed("arrow")
+
+  file <- withr::local_file(tempfile(fileext = ".parquet"))
+
+  df <- ln$core$datasets$small_dataset1()
+
+  expect_no_error(arrow::write_parquet(df, file))
+
+  expect_no_error(loaded_df <- load_file(file))
+
+  # create a Parquet file
+  df <- data.frame(a = 1:3, b = 4:6)
+  arrow::write_parquet(df, file)
 
   # load the Parquet file
   loaded_df <- load_file(file)